Audio is pauseable and adjusted normal on wall mat
