+++ Assets/Models/Environment/Walls/M_Walls.mat
@@ -13,7 +13,7 @@ Material:
   m_ModifiedSerializedProperties: 0
   m_ValidKeywords: []
   m_InvalidKeywords: []
-  m_LightmapFlags: 2
+  m_LightmapFlags: 1
   m_EnableInstancingVariants: 0
   m_DoubleSidedGI: 1
   m_CustomRenderQueue: -1
@@ -47,8 +47,8 @@ Material:
     m_Ints: []
     m_Floats:
     - _Blend: 1000
-    - _Gloss: 0
-    - _NormalStrength: 0
+    - _Gloss: -0
+    - _NormalStrength: 1
     - _QueueControl: 0
     - _QueueOffset: 0
     - _Tililng: 0.1
