+++ Assets/Prefabs/Hazards/Arena Crane.prefab
@@ -149,7 +149,7 @@ MonoBehaviour:
   containerSpawnPoint: {fileID: 4506286382292131787}
   crane: {fileID: 9054553156831217598}
   restingPoint: {fileID: 2973739573578612064}
-  alarm: {fileID: 9170956524932169260}
+  alarm: {fileID: 230246788970534632}
   creaking: {fileID: 7617132722481355205}
   craneLightFlash: {fileID: 6595181252880499882}
   dropZone: {fileID: 4320347579456167371}
@@ -933,6 +933,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   emitter: {fileID: 9170956524932169260}
+  pausable: 1
 --- !u!1001 &8849451711201945651
 PrefabInstance:
   m_ObjectHideFlags: 0
