--- Assets/Scripts/Systems/Audio/SoundVarSetter.cs
@@ -1,14 +1,0 @@
-using UnityEngine;
-
-public class SoundVarSetter : MonoBehaviour
-{
-    [SerializeField]
-    FMODUnity.StudioEventEmitter emitter;
-    FMOD.Studio.EventInstance eventInstance;
-
-    void Start()
-    {
-        eventInstance = FMODUnity.RuntimeManager.CreateInstance(emitter.EventReference);
-        FMODUnity.RuntimeManager.AttachInstanceToGameObject(eventInstance, gameObject, true);
-    }
-}
